requirements.txt
anyio==3.5.0
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.72.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
more-itertools==8.12.0
numpy==1.22.1
oauthlib==3.1.1
orjson==3.6.5
pandas==1.3.5
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
tweepy==4.4.0
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
